+++ src/dplyr.cpp
@@ -1430,6 +1430,15 @@ SEXP strip_group_attributes(DataFrame df){
 }
 
 // [[Rcpp::export]]
+DataFrame as_regular_df(DataFrame df){
+  DataFrame copy = shallow_copy(df) ;
+  SET_ATTRIB(copy, strip_group_attributes(df)) ;
+  SET_OBJECT(copy, OBJECT(df)) ;
+  copy.attr("class") = CharacterVector::create("data.frame") ;
+  return copy ;
+}
+
+// [[Rcpp::export]]
 DataFrame ungroup_grouped_df( DataFrame df){
   DataFrame copy = shallow_copy(df) ;
   SET_ATTRIB(copy, strip_group_attributes(df)) ;